+++ FalloutStrategy/FalloutEngine/AI_Player.cpp
@@ -36,10 +36,10 @@ AI_Player::AI_Player(Faction g_faction) : GenericPlayer(), is_attacking(false), 
 
 AI_Player::~AI_Player() 
 {
-	for (int t = 0; t < troops.size(); t++) { troops[t] = nullptr;}
+	for (size_t t = 0; t < troops.size(); t++) { troops[t] = nullptr;}
 	troops.clear();
 
-	for (int g = 0; g < gatherers_vector.size(); g++) { gatherers_vector[g] = nullptr; }
+	for (size_t g = 0; g < gatherers_vector.size(); g++) { gatherers_vector[g] = nullptr; }
 	gatherers_vector.clear();
 
 	target_player = nullptr;
@@ -53,7 +53,7 @@ bool AI_Player::Update(float dt) {
 	// Gather -----------------------------------------------------
 
 	if (!gatherers_commanded) {
-		for (int i = 0; i < gatherers_vector.size(); i++)
+		for (size_t i = 0; i < gatherers_vector.size(); i++)
 		{
 			//authomatic gathering
 			if (gatherers_vector[i]->GetResourceBuilding() == nullptr) {
@@ -137,7 +137,7 @@ bool AI_Player::Update(float dt) {
 				target_building_position = target_building->current_tile;
 		}
 
-		for (int i = 0; i < troops.size(); i++)
+		for (size_t i = 0; i < troops.size(); i++)
 		{
 			troops[i]->target_building = target_building;
 		}
@@ -166,7 +166,7 @@ DynamicEntity* AI_Player::GetClosestDynamicEntity() {
 	DynamicEntity* target_entity = nullptr;
 	int distance = 1000000;
 
-	for (int i = 0; i < target_player->troops.size(); i++)
+	for (size_t i = 0; i < target_player->troops.size(); i++)
 	{
 		if ((target_player->troops[i]->current_tile.DistanceManhattan(base->current_tile) < distance)&&(target_player->troops[i])) {
 			target_entity = target_player->troops[i];
@@ -194,7 +194,7 @@ StaticEntity* AI_Player::ChooseTargetBuilding() {
 
 void AI_Player::GatherFood(ResourceBuilding* resource_spot) {
 
-	for (int i = 0; i < gatherers; i++)
+	for(size_t i = 0; i < gatherers; i++)
 	{
 		if ((gatherers_vector[i]->GetResourceBuilding() == nullptr)||(gatherers_vector[i]->GetResourceCollected() == 0)) {
 			gatherers_vector[i]->AssignResourceBuilding(resource_spot);
